requirements.txt
# Core
requests>=2.31.0
aiohttp>=3.9.0
runpod>=1.6.0
pillow>=10.0.0
numpy>=1.24.0

# Image processing (GPU)
rembg[gpu]>=2.0.50
onnxruntime-gpu>=1.16.0
realesrgan>=0.3.0
basicsr>=1.4.2
torch>=2.0.0

# LangGraph (swarm orchestration)
langgraph>=0.2.0
langchain-core>=0.3.0

# SearXNG (bundled)
searxng>=2024.1.0
uvloop>=0.19.0
httpx[http2]>=0.27.0
